pyproject.toml
[tool.poetry]
name = "my-app"
version = "0.1.0"
description = ""
authors = ["cs316"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^2.2"
djangorestframework = "^3.10"
whitenoise = "^4.1"
gunicorn = "^19.9"
django-heroku = "^0.3.1"
django-admin = "^2.0"
firebase-admin = "^3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
